Add Time for Last Restart (#25)
We now display the time, when the Pod was restarted as last. Instead of
having two variables (`hasRecentRestarts` and `restartAge`) we are just
using the `restartAge` to decide if the Pod should be diplayed because
of the restarts or not. This is possible because the string will be
empty if it hadn't recent restarts.